added endpoints get of user and profiles
