+++ InterLab.API/InterLab.API/Controllers/CompanyInternshipsController.cs
@@ -47,6 +47,44 @@ namespace InterLab.API.Controllers
             return Ok(internshipResource);
         }
 
+        [HttpPost]
+        public async Task<IActionResult> PostAsync([FromBody] SaveInternshipResource resource)
+        {
+            if (ModelState.IsValid)
+                return BadRequest(ModelState.GetErrorMessages());
+            var internship = _mapper.Map<SaveInternshipResource, Internship>(resource);
+            var result = await _internshipService.SaveAsync(internship);
 
+            if (!result.Success)
+                return BadRequest(result.Message);
+
+            var internshipResource = _mapper.Map<Internship, InternshipResource>(result.Resource);
+            return Ok(internshipResource);
+        }
+
+        [HttpPut("{id}")]
+        public async Task<IActionResult> PutAsync(int id, [FromBody] SaveInternshipResource resource)
+        {
+            var internship = _mapper.Map<SaveInternshipResource, Internship>(resource);
+            var result = await _internshipService.UpdateAsync(id, internship);
+
+            if (!result.Success)
+                return BadRequest(result.Message);
+            var internshipResource = _mapper.Map<Internship, InternshipResource>(result.Resource);
+            return Ok(internshipResource);
+
+
+        }
+
+        [HttpDelete("{id}")]
+        public async Task<IActionResult> DeleteAsync(int id)
+        {
+            var result = await _internshipService.DeleteAsync(id);
+
+            if (!result.Success)
+                return BadRequest(result.Message);
+            var internshipResource = _mapper.Map<Internship, InternshipResource>(result.Resource);
+            return Ok(internshipResource);
+        }
     }
 }
