+++ InterLab.API/InterLab.API/Controllers/ProfilesController.cs
@@ -1,4 +1,5 @@
 using AutoMapper;
+using InterLab.API.Domain.Models;
 using InterLab.API.Domain.Services;
 using InterLab.API.Extensions;
 using InterLab.API.Resources;
@@ -10,7 +11,7 @@ using System.Threading.Tasks;
 namespace InterLab.API.Controllers
 {
 
-    [Route("/api/users/{userId}/profiles")]
+    [Route("/api/profiles")]
     public class ProfilesController : Controller
     {
         private readonly IProfileService _profileService; 
@@ -22,45 +23,22 @@ namespace InterLab.API.Controllers
             _mapper = mapper;
         }
 
-        [HttpPost]
-        public async Task<IActionResult> PostAsync([FromBody] SaveProfileResource resource)
+        [HttpGet]
+        public async Task<IEnumerable<ProfileResource>> GetAllAsync()
         {
-            if (ModelState.IsValid)
-                return BadRequest(ModelState.GetErrorMessages());
-            var profile = _mapper.Map<SaveProfileResource, Domain.Models.Profile>(resource);
-            var result = await _profileService.SaveAsync(profile);
-
-            if (!result.Success)
-                return BadRequest(result.Message);
-
-            var profileResources = _mapper.Map<Domain.Models.Profile, ProfileResource>(result.Resource);
-            return Ok(profileResources);
+            var profiles = await _profileService.ListAsync();
+            var resources = _mapper.Map<IEnumerable<Domain.Models.Profile>, IEnumerable<ProfileResource>>(profiles);
+            return resources;
         }
 
-        [HttpDelete("{id}")]
-        public async Task<IActionResult> DeleteAsync(int id)
+        [HttpGet("{id}")]
+        public async Task<IActionResult> GetAsync(int id)
         {
-            var result = await _profileService.DeleteAsync(id);
-
+            var result = await _profileService.GetByIdAsync(id);
             if (!result.Success)
                 return BadRequest(result.Message);
             var profileResource = _mapper.Map<Domain.Models.Profile, ProfileResource>(result.Resource);
             return Ok(profileResource);
-        }
-
-
-        [HttpPut("{id}")]
-        public async Task<IActionResult> PutAsync(int id, [FromBody] SaveProfileResource resource)
-        {
-            var profiles = _mapper.Map<SaveProfileResource, Domain.Models.Profile>(resource);
-            var result = await _profileService.UpdateAsync(id, profiles);
-
-            if (!result.Success)
-                return BadRequest(result.Message);
-            var profileResource = _mapper.Map<Domain.Models.Profile, ProfileResource>(result.Resource);
-            return Ok(profileResource);
-
-
         }
     }
