+++ InterLab.API/InterLab.API/Controllers/UsersController.cs
@@ -32,6 +32,15 @@ namespace InterLab.API.Controllers
             return resources;
         }
 
+        [HttpGet("{id}")]
+        public async Task<IActionResult> GetAsync(int id)
+        {
+            var result = await _userService.GetById(id);
+            if (!result.Success)
+                return BadRequest(result.Message);
+            var profileResource = _mapper.Map<User, UserResource>(result.Resource);
+            return Ok(profileResource);
+        }
 
     }
