+++ InterLab.API/InterLab.API/appsettings.json
@@ -7,7 +7,7 @@
     }
   },
   "ConnectionStrings": {
-    "DefaultConnection": "server=localhost;port=3306;database=interlab_c#;uid=root;password=jchcjuan0312"
+    "DefaultConnection": "server=localhost;port=3306;database=interlab_c#;uid=root;password=#8rbStudios"
   },
   "AllowedHosts": "*"
 }
